+++ CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 ### Added
 1. New commands `chronometrist-restart-task`, `chronometrist-extend-task`
-2. Menu of commands available in `chronometrist`
+2. Menus for `chronometrist` and `chronometrist-key-values`
 ### Changed
 3. Display graph ranges in `chronometrist-spark` column
+4. `chronometrist-tags-add` and `chronometrist-key-values-unified-prompt` now also work interactively.
 
 ## [0.8.1] - 2021-06-01
 ### Changed